+++ Module2/src/case_study/FuramaResort/model/Room.java
@@ -3,7 +3,7 @@ package case_study.FuramaResort.model;
 public class Room extends Facility {
     private String freeService;
 
-    public Room(String facilityCode, String facilityName, int area, int cost, int maxPeople, String typeHire, String freeService) {
+    public Room(String facilityCode, String facilityName, double area, int cost, int maxPeople, String typeHire, String freeService) {
         super(facilityCode, facilityName, area, cost, maxPeople, typeHire);
         this.freeService = freeService;
     }
